Update scripts
Final update for the night. With this I no longer have to do edits to the repo to compile. I just have to export a few environement variables related to my setup. This will also stop me from doing partial pushes and all that silly stuff, keeping me with the HEAD.